separate archive and update scripts
