+++ faq-site/package.json
@@ -7,8 +7,9 @@
         "clean": "npx rimraf _site",
         "build": "npx @11ty/eleventy",
         "serve": "npx @11ty/eleventy --serve",
-        "update-data": "npx cross-env AZURE_ENVIRONMENT=test node tools/updateData.js",
-        "lint": "node tools/lintData.js",
+        "archive-data": "node tools/archiveData.js",
+        "fetch-data": "npx cross-env AZURE_ENVIRONMENT=test node tools/updateData.js",
+        "lint-data": "node tools/lintData.js",
         "restore-local": "npx cross-env AZURE_ENVIRONMENT=test node tools/restoreTestKb.js",
         "deploy-to-prod": "node tools/deployToProd.js"
     },
