+++ faq-site/tools/updateData.js
@@ -1,6 +1,5 @@
-const { promises: fs } = require("fs");
 const qnaMakerApi = require('@ads-vdh/qnamaker-api');
-const GENERATED_FILE_WARNING = "// GENERATED FILE - only update by re-running updateData.js - local changes will be wiped out\r\n"
+const { writeFile } = require('./utilities')
 
 let filePath = process.env.AZURE_ENVIRONMENT ? `.env.${process.env.AZURE_ENVIRONMENT}` : ".env"
 require('dotenv').config({ path: filePath })
@@ -16,20 +15,9 @@ async function updateData() {
         kbId: process.env.kbId
     })
 
-    await archiveFaqs()
     await updateKnowledgeBase(client)
     await updateAlterations(client)
 
-}
-
-async function archiveFaqs() {
-    const { readJsonc } = require('./utilities')
-
-    let faqs = await readJsonc("_data/faqs.jsonc")
-
-    let contents = JSON.stringify(faqs, null, 4);
-
-    await writeFile("_data/faqs-prev.jsonc", contents)
 }
 
 async function updateKnowledgeBase(client) {
@@ -63,19 +51,4 @@ async function updateAlterations(client) {
     let contents = "var synonyms = " + JSON.stringify(synonyms, null, 4);
 
     await writeFile("assets/synonyms.js", contents)
-}
-
-async function writeFile(path, contents) {
-
-    let projectRoot = __dirname.replace(/tools$/, "");
-    let fullPath = `${projectRoot}/${path}`;
-
-    try {
-        await fs.writeFile(fullPath, GENERATED_FILE_WARNING + contents)
-
-        console.log(`Data has been written to ${path}`);
-
-    } catch (error) {
-        console.error(error)
-    }
 }
